+++ include/group.hpp
@@ -11,8 +11,8 @@ class Group : public Object3D {
 
   ~Group() override {}
 
-  bool intersect(const Ray &r, Hit &h, Object3D *&obj, float tmin) override {
-    float _tmin = tmin, _tmax = h.t;
+  bool intersect(const Ray &r, Hit &h, Object3D *&obj, double tmin) override {
+    double _tmin = tmin, _tmax = h.t;
     if (!bbox.intersect(r, _tmin, _tmax)) return false;
     bool result = false;
     for (Object3D *object : objects)
